add store, add SearchParameter component
