+++ src/reducers/index.js
@@ -1,0 +1,9 @@
+
+
+const initialState = {
+    searchParameters: "banana"
+  };
+
+  const rootReducer = (state = initialState, action) => state;
+
+  export default rootReducer;
