+++ src/store/index.js
@@ -1,0 +1,7 @@
+import { createStore } from "redux";
+import rootReducer from "../reducers/index";
+
+const store = createStore(rootReducer);
+window.store = store;
+
+export default store;
